requirements.txt
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy==1.3.20
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
#googletrans
google_trans_new
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.16.4
spamwatch
alphabet_detector
pyrate-limiter
cachetools
fontTools
gtts
pyrogram
pendulum
wget
Faker
Python_arq
redis
nekos.py
tswift
pyowm
tabulate
parsel
pyrogram
tgcrypto
telegraph
pytz
hachoir
pybase64
pySmartDL
validators
youtube_dl
html2text
bing_image_downloader
register
dispatcher
cloudscraper
img2pdf
nhentai.py
heroku3
opencv-python
youtube_dl
mtranslate
envparse==0.1.6
PyDictionary
search-engine-parser
youtube-search-python
countryinfo
motor
Python_ARQ
